+++ Flow/Changes.md
@@ -9,6 +9,9 @@
   - Added `.groupedWithinSize(Int)` for deep collections
 - **Seq** (**CollectionExtensions**)
   - Added `.tryReduce(...)` and `.tryReduceIterator(...)`
+### Other changes
+- **ValueConvertible**'s `.toValue` is no longer implicit
+- **UncertainBoolean** now extends **ValueConvertible**
 
 ## v2.6 - 26.05.2025
 This update focuses on improving the **Changing** interface internally, and by adding lockable flags.